WebServlet for Struts

+++ web-servlet/src/main/webapp/WEB-INF/views/person.jsp
@@ -1,7 +1,9 @@
 <%@page pageEncoding="utf-8" %>
 <%@taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
+<%@taglib uri="http://java.sun.com/jsp/jstl/fmt" prefix="fmt" %>
 <html>
 <head>
+  <link rel="stylsheet" type="text/css" href="../css/style.css">
   <meta charset="utf-8" />
   <title></title>
 </head>
@@ -23,7 +25,7 @@
     <c:if test="${empty id}">
       <tr>
         <td>*</td>
-        <td><input name="firstName" /></td>
+        <td><input class="" name="firstName" /></td>
         <td><input name="lastName" /></td>
         <td><input name="email" /></td>
         <td><input name="birthDate" /></td>
@@ -37,8 +39,9 @@
           <td><input name="firstName" value="${person.firstName}"></td>
           <td><input name="lastName" value="${person.lastName}"></td>
           <td><input name="email" value="${person.email}"></td>
-<%--          <td><input name="birthDate" value="${person.birthDate}"></td>--%>
-          <td><a href="${context}delete/${person.id}">Supprimer</a></td>
+          <fmt:parseDate value="${person.birthDate}" pattern="yyyy-MM-dd" var="parsedDate" type="date"/>
+          <fmt:formatDate value="${parsedDate}" type="date" pattern="dd/MM/yyyy" var="goodDate"/>
+          <td><input name="birthDate" value="${goodDate}"></td>
           <td><a href="${context}">Annuler</a></td>
           <td><input type="submit" /></td>
         </tr>
@@ -50,7 +53,7 @@
           <td>${person.lastName}</td>
           <td>${person.email}</td>
           <td>${person.birthDate}</td>
-          <td><a href="${context}delete/${person.id}">Supprimer</a></td>
+          <td><a href="${context}delete/${person.id}" onclick="return confirm('Etes-vous sur de vouloir supprimer ?')">Supprimer</a></td>
           <td><a href="${context}${person.id}">Modifier</a></td>
         </tr>
       </c:if>
@@ -58,11 +61,6 @@
     </tbody>
   </table>
 </form>
-<%--<script>--%>
-<%--  function edit(event){--%>
-<%--    console.log(event);--%>
-<%--    event.preventDefault();--%>
-<%--  }--%>
-<%--</script>--%>
+
 </body>
 </html>
